requirements.txt
easyocr==1.7.1
fastapi==0.111.1
openai==1.36.1
pydantic==2.8.2
python-bidi==0.4.2
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.23.0
torch==2.0.0
uvicorn==0.30.3